Finish `./ugit init` command
Now, we use `CLI11` library to parse the command. It is stupid to hold
the commands into only one file. So I decide to divide them. Although
this is a small project. But the idea is important. Now the `./ugit init`
command just echo "Hello, World!\n", I will handle this in the next
commit