feature: add `./ugit hash-object` command
Actually, the idea is simple. This commit uses OpenSSL library's sha1sum
function to generate has id. However, things are trivial when using cpp.
We need to convert the binary string into hex string. It may seem that
we could use `printf` or `cout` to generate the hex string. Well, its
efficiency is low. We need to use lookup table method. It is easy. We
calculate for every 4 bit.

Next, we use cpp file stream to write and read files. Well, it is really
tedious. But the code I design now is clear. I think this is worthy.