Finish `./ugit write-tree` and `./ugit read-tree` command
Now we have second type "tree". The tree is just like a directory. So
what we do first is to recursively iterate the directory. When
encountering a file, we record the information into a tuple. And when
all done, we write a tree object to contain the information of the
current directory.

The `read-tree` is reverse. Use C++ to implement this is a little
trivial. Be careful.