feature: finish `./ugit commit -m (--message)` command
Now we have created the thrid type in the `ugit`—`commit`. Actually
commit stores the top-level tree object ID. And some message we writes.
In order to indicate where we are. We write the current commit object ID
to the `HEAD`. And in order to implement the `ugit log` we need to
stores the last commit ID in the current commit.

fixes some spell errors