feature: Add tag resolve
Now we have already created the tag which is a readable name to
represent hash object id. Now we need to resolve the tag name. It is not
difficult.

Also, this commit refactors the code. In order to make the code more
readable, this commit makes the variable name more concrete. As you can
see, there are three types for the object:

+ blob
+ tree
+ object

Tag could be name alias for any objects.