feature: finish branch features
Well, it is difficult to finish the branch. Branch actually is just a
name which stores the current commit ID. And the `HEAD` is a symbolic
file which points to the branch files.

So this commit just handles this. However, I think the current code is
not elegant. So need to refactor.

TODO: refactor